pyhat sbs - native module

+++ src_c/hat/sbs.h
@@ -23,7 +23,7 @@ size_t hat_sbs_encode_union_header(hat_buff_t *buff, size_t id);
 
 int hat_sbs_decode_boolean(hat_buff_t *buff, _Bool *value);
 int hat_sbs_decode_integer(hat_buff_t *buff, int64_t *value);
-int hat_sbs_decode_float(hat_buff_t *buff, float *value);
+int hat_sbs_decode_float(hat_buff_t *buff, double *value);
 int hat_sbs_decode_string(hat_buff_t *buff, uint8_t **value, size_t *value_len);
 int hat_sbs_decode_bytes(hat_buff_t *buff, uint8_t **value, size_t *value_len);
 int hat_sbs_decode_array_header(hat_buff_t *buff, size_t *len);
